requirements.txt
pandas
scikit-learn
bentoml
flask
pyjwt
pytest
fastapi